Rename package from pay to payments
