+++ example/src/androidTest/java/org/wordpress/android/fluxc/mocked/MockedStack_WCPayTest.kt
@@ -6,14 +6,14 @@ import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
 import org.junit.Test
 import org.wordpress.android.fluxc.model.SiteModel
-import org.wordpress.android.fluxc.model.pay.WCCapturePaymentErrorType.CAPTURE_ERROR
-import org.wordpress.android.fluxc.model.pay.WCCapturePaymentErrorType.MISSING_ORDER
-import org.wordpress.android.fluxc.model.pay.WCCapturePaymentErrorType.PAYMENT_ALREADY_CAPTURED
-import org.wordpress.android.fluxc.model.pay.WCCapturePaymentErrorType.SERVER_ERROR
-import org.wordpress.android.fluxc.model.pay.WCPaymentAccountResult.WCPayAccountStatusEnum
-import org.wordpress.android.fluxc.model.pay.WCTerminalStoreLocationErrorType.GenericError
-import org.wordpress.android.fluxc.model.pay.WCTerminalStoreLocationErrorType.InvalidPostalCode
-import org.wordpress.android.fluxc.model.pay.WCTerminalStoreLocationErrorType.MissingAddress
+import org.wordpress.android.fluxc.model.payments.WCCapturePaymentErrorType.CAPTURE_ERROR
+import org.wordpress.android.fluxc.model.payments.WCCapturePaymentErrorType.MISSING_ORDER
+import org.wordpress.android.fluxc.model.payments.WCCapturePaymentErrorType.PAYMENT_ALREADY_CAPTURED
+import org.wordpress.android.fluxc.model.payments.WCCapturePaymentErrorType.SERVER_ERROR
+import org.wordpress.android.fluxc.model.payments.WCPaymentAccountResult.WCPayAccountStatusEnum
+import org.wordpress.android.fluxc.model.payments.WCTerminalStoreLocationErrorType.GenericError
+import org.wordpress.android.fluxc.model.payments.WCTerminalStoreLocationErrorType.InvalidPostalCode
+import org.wordpress.android.fluxc.model.payments.WCTerminalStoreLocationErrorType.MissingAddress
 import org.wordpress.android.fluxc.module.ResponseMockingInterceptor
 import org.wordpress.android.fluxc.network.rest.wpcom.wc.pay.PayRestClient
 import javax.inject.Inject
