+++ plugins/woocommerce/src/main/kotlin/org/wordpress/android/fluxc/model/payments/WCConnectionTokenResult.kt
@@ -1,3 +1,3 @@
-package org.wordpress.android.fluxc.model.pay
+package org.wordpress.android.fluxc.model.payments
 
 data class WCConnectionTokenResult(val token: String, val isTestMode: Boolean)
