+++ plugins/woocommerce/src/main/kotlin/org/wordpress/android/fluxc/network/rest/wpcom/wc/pay/PayRestClient.kt
@@ -5,20 +5,20 @@ import com.android.volley.RequestQueue
 import org.wordpress.android.fluxc.Dispatcher
 import org.wordpress.android.fluxc.generated.endpoint.WOOCOMMERCE
 import org.wordpress.android.fluxc.model.SiteModel
-import org.wordpress.android.fluxc.model.pay.WCCapturePaymentError
-import org.wordpress.android.fluxc.model.pay.WCCapturePaymentErrorType.CAPTURE_ERROR
-import org.wordpress.android.fluxc.model.pay.WCCapturePaymentErrorType.GENERIC_ERROR
-import org.wordpress.android.fluxc.model.pay.WCCapturePaymentErrorType.MISSING_ORDER
-import org.wordpress.android.fluxc.model.pay.WCCapturePaymentErrorType.NETWORK_ERROR
-import org.wordpress.android.fluxc.model.pay.WCCapturePaymentErrorType.PAYMENT_ALREADY_CAPTURED
-import org.wordpress.android.fluxc.model.pay.WCCapturePaymentErrorType.SERVER_ERROR
-import org.wordpress.android.fluxc.model.pay.WCCapturePaymentResponsePayload
-import org.wordpress.android.fluxc.model.pay.WCPaymentAccountResult
-import org.wordpress.android.fluxc.model.pay.WCPaymentCreateCustomerByOrderIdResult
-import org.wordpress.android.fluxc.model.pay.WCTerminalStoreLocationError
-import org.wordpress.android.fluxc.model.pay.WCTerminalStoreLocationErrorType
-import org.wordpress.android.fluxc.model.pay.WCTerminalStoreLocationResult
-import org.wordpress.android.fluxc.model.pay.WCTerminalStoreLocationResult.StoreAddress
+import org.wordpress.android.fluxc.model.payments.WCCapturePaymentError
+import org.wordpress.android.fluxc.model.payments.WCCapturePaymentErrorType.CAPTURE_ERROR
+import org.wordpress.android.fluxc.model.payments.WCCapturePaymentErrorType.GENERIC_ERROR
+import org.wordpress.android.fluxc.model.payments.WCCapturePaymentErrorType.MISSING_ORDER
+import org.wordpress.android.fluxc.model.payments.WCCapturePaymentErrorType.NETWORK_ERROR
+import org.wordpress.android.fluxc.model.payments.WCCapturePaymentErrorType.PAYMENT_ALREADY_CAPTURED
+import org.wordpress.android.fluxc.model.payments.WCCapturePaymentErrorType.SERVER_ERROR
+import org.wordpress.android.fluxc.model.payments.WCCapturePaymentResponsePayload
+import org.wordpress.android.fluxc.model.payments.WCPaymentAccountResult
+import org.wordpress.android.fluxc.model.payments.WCPaymentCreateCustomerByOrderIdResult
+import org.wordpress.android.fluxc.model.payments.WCTerminalStoreLocationError
+import org.wordpress.android.fluxc.model.payments.WCTerminalStoreLocationErrorType
+import org.wordpress.android.fluxc.model.payments.WCTerminalStoreLocationResult
+import org.wordpress.android.fluxc.model.payments.WCTerminalStoreLocationResult.StoreAddress
 import org.wordpress.android.fluxc.network.BaseRequest.GenericErrorType
 import org.wordpress.android.fluxc.network.UserAgent
 import org.wordpress.android.fluxc.network.rest.wpcom.BaseWPComRestClient
